added Hackerrank Sherlock and Anagrams.

+++ src/test/java/com/xudong/algo/array/PrefixSumAndDifferenceTest.java
@@ -18,7 +18,7 @@ public class PrefixSumAndDifferenceTest {
         int n = 10;
 
         //when
-        int biggest  = new PrefixSumAndDifference().manipulateAndReturnBiggest(n, manipulations);
+        long biggest  = new PrefixSumAndDifference().manipulateAndReturnBiggest(n, manipulations);
 
         //then
         assertThat(biggest).isEqualTo(10);
